+++ test/linklist_test.c
@@ -4,11 +4,11 @@
 #include <testing.h>
 #include <linklist.h>
 
-int iterator_callback(void *item, unsigned long idx, void *user) {
+int iterator_callback(void *item, uint32_t idx, void *user) {
     int *failed = (int *)user;
     char *val = (char *)item;
     char test[100];
-    sprintf(test, "test%lu", idx+1);
+    sprintf(test, "test%u", idx+1);
     if (strcmp(test, val) != 0) {
         t_failure("Value at index %d doesn't match %s  (%s)", idx, test, val);
         *failed = 1;
